Add files via upload
I built a model of housing prices in California using the California census data. This data has metrics such as the population, median income, median housing price, and so on for each block group in California. 
Block groups are the smallest geographical unit for which the US Census Bureau publishes sample data (a block group typically has a population of 600 to 3,000 people).